Set dm ioctl structs with 0-len members to packed
In many cases, the dm ioctls use structs with 0-size final members to refer
to a variable-length array, such as a string. We do not want Rust to
round the struct size up to its natural alignment, because then we get
the wrong answer when trying to calculate the last members' offset. Use
the 'packed' directive to achieve this.

However, we seem to be in shaky territory here because I found that
setting Struct_dm_ioctl to packed caused problems. There's more
investigation needed here.

Signed-off-by: Andy Grover <agrover@redhat.com>