Add LinearDev::extend()
Allow a LinearDev to be extended by passing in more Segments.

Remember in-use segments so we can properly re-load the table when adding
new ones.

extend() assumes ::new() is called with a non-empty vec of Segments. Add an
assert, and also in dm_table().

Try to avoid multiple segments representing a single contiguous range.

Make members of Segment public, to enable this. I think this is OK
because Segment is used more as a container for commonly-used-together
variables, rather than something that is presenting a clean API over
something that is complex.

Signed-off-by: Andy Grover <agrover@redhat.com>