Change DmError::Dm constructor definition
Have it take an ErrorEnum and a String as argument.

Update implementation and invocations accordingly.

This brings our approach to error handling into line with what we have
come up with in stratisd and cleans up some dead code.

Previously ErrorEnum was unused and Internal was used; now Internal is gone
and ErrorEnum is used.

Signed-off-by: mulhern <amulhern@redhat.com>