Add a Makefile target that runs only non-root tests
It is convenient to run basic unit tests and non-running doc tests
on one's own system.

On the CI, continue to run all tests as root.

This is a somewhat temporary expedient. If more tests that require root
are added, it may pay to better distinguish between the different types of
tests.

Signed-off-by: mulhern <amulhern@redhat.com>