Return data & meta lineardevs from ThinPoolDev::setup if CheckFailed
Since these are moved into the function, if it fails then they are dropped
and the caller can't use the LinearDevs any more. Instead, move them back
out if the check fails, so the caller can attempt to recover.

To this end, remove Clone from ErrorEnum, since LinearDev is not Clone.

Signed-off-by: Andy Grover <agrover@redhat.com>