Make Device::devnode return a Result type
Previously it returned an Option<PathBuf>, now it returns
DmResult<Option<PathBuf>>. This is more correct, because it allows to
distinguish between an error and the device simply not having a devnode,
a situation which does occur.

Even if this method is changed so that it no longer reads from
/proc/partitions the change in its return type is correct, as it must
obtain the devnode somehow from the environment.

The callers continue to return an error if there is no devnode, since it
is assumed that all devicemapper devices have a devnode.

Signed-off-by: mulhern <amulhern@redhat.com>