Establish a shared module for functionality shared by DM targets
This kind of functionality does not belong with the core DM implementation.
It is not low-level like the code in the utils module.
It is shared, and these methods are the start of a more properly shared
set of methods implementing DM targets.

Implement the methods table_load(), table_reload(), and device_exists().

Remove DM::table_reload(), since it has been transferred to shared module.

Switch to using shared table_load() or table_reload() where appropriate

Do not yet use device_exists().

Signed-off-by: mulhern <amulhern@redhat.com>