Change new() or setup() method for greater idempotency
Avoid failing if the cause of the failure is that the device is already
known to the kernel.

Postpone verifying if the arguments match the kernel's model to a separate
PR. It is essential that this be done, but it is a bit of a task, and, under
normal operation, it is unnecessary.

device_exists() is no longer dead code, so remove the annotation.

Signed-off-by: mulhern <amulhern@redhat.com>