Fix remaining or_fun_call clippy errors
These are all to be found in devnode() methods. At one time, we thought
we had an improvement to the devnode() methods, whch as a side-effect
removed the error. But that improvement was reverted and so the errors
returned. The associated issue is:
https://github.com/stratis-storage/devicemapper-rs/issues/106.
Unfortunately, that issue is kind of tricky, and may not be fixed very
soon. It would be nice to make the clippy target non-optional, and these
are the only errors remaining. Therefore, I'm fixing these errors,
disregarding the fact that the functions remaing unfixed.

Signed-off-by: mulhern <amulhern@redhat.com>