Rewrite device_rename() to take DevId target argument
Use the constructor of the target argument to decide whether to pass
DM_UUID flag.

The purpose is to gain some correctness by avoiding mixing up names and
devicemapper uuids. There is only some safety gained, because a DM
uuid is not a UUID, so it is impossibly to use the Uuid type.

This fixes a bug, since the DM_UUID flag does not specify that the uuid
should be reset, but specifyies that a uuid may be set if one has not
been set already.

Signed-off-by: mulhern <amulhern@redhat.com>