Move FromStr for Device into a utility function
Omit the parsing of the device number from a string, it seems unused
and it makes the interface uglier.

Make the utility function, devnode_to_devno(), part of the testing
infrastructure. It is unused in source, but handy for testing.
Change what were previously error returns into unwraps, for conciseness.

Signed-off-by: mulhern <amulhern@redhat.com>