Rewrite DevId type
Implement DevId in terms of new types DmUuid and DmName. These types were
previously aliases for str, now they are complete types. The types give
type safety and ensure that their values are ASCII and do not exceed the
maximum length allowed.

Make DevId Copy. It doesn't need to be passed as a reference, because each
value is just a tag indicating Name or Uuid and a reference to a str value.

Signed-off-by: mulhern <amulhern@redhat.com>