Make DmName/DmUuid unsized types
This changes DmName from a type that *has* a reference into one that *is*
a reference. This increases the similarity in use and implementation with
analogous structs in the stdlib such as Path/PathBuf, and makes
implementing ToOwned trait (for the ref type) and Borrow (for the owned
type) straightforward.

Implement ToOwned.

Change macro $T/$B idents to $B/$O, this is more clearly delineates between
borrowed and owned ident usage.

The only unpleasantness is the need for transmute() in $B::new. I looked
in the Rust stdlib and this is also used by CStr, OsStr and Path, so it
appears unavoidable.

This commit also uses a single macro to define both the borrowed and owned
types.

Move definition of borrowed and owned structs inside the dev_id macro.

Change device_exists() implementation to use into_iter() to simplify
comparison in the any() predicate.

Add more comments.

It's not clear to me why we need as_ref in a few places when we didn't when
using str/String, add a TODO.

Signed-off-by: Andy Grover <agrover@redhat.com>