Tidy some mentions of Stratis in devicemapper-rs
Devicemapper stands on its own. There are a few harmless mentions of
Stratis in the code and test strings, but let's try to avoid these just on
principle.

Take the opportunity to describe in more detail how we can be sure the DM
kernel interface won't change under us and cause a crash.

Signed-off-by: Andy Grover <agrover@redhat.com>