Return an error if a DmUuid or DmName value has zero characters
The empty string is used to mean None, and is treated that way by DM.
For example, if you try to make a device with an empty string as the name,
an error is returned. No uuid is represented by an empty string in the
dm_info struct, and so forth.

Signed-off-by: mulhern <amulhern@redhat.com>