Disperse most of the constants defined in consts.rs into other modules
Many are explicitly associated with one module, and might as well be
localized there. In C, we couldn't do this because we couldn't have
constants in mutually dependent modules, but in Rust there is no such
problem.

Some constants are associated with one module; these are private, for
better encapsulation.

Signed-off-by: mulhern <amulhern@redhat.com>