Do not run thin_check as part of thinpool setup
Expose device_exists() function for the convenience of client code.

Reasons:
1. devicemapper is a fairly low level library. It is generally a good idea
for libraries not to shell out to command-line tools, for low-level
libraries this is doubly true.
2. It is non-optional. Users who are confident that their metadata is in
good repair, will not be pleased that they are shelling out to a
command-line tool for no reason.
3. It makes calling ThinPoolDev::setup() from client code rather awkward.
It would be better for the client just to run thin_check directly on their
metadata device, and take action as necessary, rather than react to a
failure in setting up the thinpool. The client can use device_exists() to
verify that it is "safe" to run thin_check. The client can decide whether it
wants to proceed w/out information, i.e., if there was an error in calling
thin_check command more readily.
4. Calling thin_check in setup() makes error generation unnecessarily
awkward. There are lots of situations that we might want to potentially
discriminate among for the benefit of the client, see (4), better to let
the client do it directly.

Example of thin_check execution in stratisd.

```
let pool name = format_thinpool_name(...);
if !device_exists(dm, name)? {
  let result = match Command::new("thin_check").... {
    Ok(result) => result,
    Err(err) {
      // client code can treat this as hard or soft error as appropriate
    }
  };
  let (meta_dev, spare_segments) = if !result.success() {
    match attempt_thin_repair(pool_uuid, dm, meta_dev, spare_segments) {
      Ok(result) => result,
      Err(err) => // treat this as hard error
    }
  };
}

// proceed to setup thinpool with verified metadata device
```

Signed-off-by: mulhern <amulhern@redhat.com>