Omit debug! in LinearDev::dm_table
I believe that it has served its purpose as a development aid, and it
would be unlikely to be all that helpful for customer debugging.

When there is a failure, we would be better served by packaging that
information up in the error returned.

Get rid of log, env_logger requirements. They can be re-added w/ a small
amount of effort as necessary.

Signed-off-by: mulhern <amulhern@redhat.com>