Add DM::into_file()
For using the new DM event mechanism in kernel 4.13, we need to get an
open fd to /dev/mapper/control. It seems like the best way is to move the
File out of a DM, and then call as_raw_fd(). We can't just implement
as_raw_fd() for DM because the file needs to remain open (i.e. in scope)
while we're polling its fd.

Signed-off-by: Andy Grover <agrover@redhat.com>