device: Convert to/from dev_t and kdev_t
We can't assume which bits in a libc::dev_t are major and which are minor.
This is especially true since the default is to use an intermingled
pattern that allows for >256 major and minor, but keeps the layout the same
as the old 16 bit dev_t if both values are less than 256.

Also properly convert to/from the kernel's dev_t encoding. Even though
DM's ioctl struct has a 64 bit field for this, it uses the kernel's
"huge" encoding, which is only 32 bits. Therefore it's possible to have
a valid Device that cannot be expressed to the kernel. But, this field
isn't currently ever set right now. If we ever want to specify DM devices
by maj:min then we'd need to use to_kdev_t().

fixes #147

Signed-off-by: Andy Grover <agrover@redhat.com>