Generate bindings using rust-bindgen as part of build
This changes from using the one-time output of rust-bindgen to generate
DM ioctl struct definitions, to generating it as part of the build process.
This keeps the definitions updated if the header file changes.

It's possible to customize generation as needed by configuring the builder
in build.rs.

The only tricky bit is that if one needs to see the generated definitions,
they now are stuck deep in target/, in a file called bindings.rs.