Verify that data block size is within bounds in ThinPoolDev methods
It's not costly to do, and it uses some otherwise dead code.

It's a lot clearer and more specific than the ioctl return code which is
what would otherwise be returned on bad values.

Export the minimum and maximum values that we believe in as well as the
verification function. These won't hurt if anybody ever wants to write a
GUI and vet the users choice.

Signed-off-by: mulhern <amulhern@redhat.com>