Rewrite ThinPoolDev status() method
Have it read more values from the data and also simplify its structure.

Also, rewrite ThinPoolStatus type so status() method can return more of
the information that it reads.

* Change Good ThinPoolStatus constructor to Working
It matches the rest of the nomenclature, and out-of-space is not really
"good".
* Have Working constructor take a single struct with a number of fields.
* Do not derive Copy on some structs that are kind of big.
* Export newly created status types in public interface.
* Replace block_usage/BlockUsage with usage/Usage
Our types gives us the necessary information, so it is not necessary to
put it in the variable name as it would be with C.

Signed-off-by: mulhern <amulhern@redhat.com>