Unparameterize DM::table_load()
It is much better to have the params type be a String, because:

1. Then it is not required to have an implementation of TargetParams,
for a particular dm device. The function can be used w/out that constraint.
2. There will not be multiple copies of the method for every implementation
of a DM device.

The downside is that conversion must occur, but usually that will be
cheap as generally there is only one line in the table.

Signed-off-by: mulhern <amulhern@redhat.com>