Add new LinearDevTargetParams type which is a union of Linear and Flakey
The params in a LinearDev table can be either LinearTargetParams or
FlakeyTargetParams.

No longer require that TargetParams implement FromStr. It is not useful
for LinearDevTargetParams to implement FromStr, because it can not parse
the unknown type directly. It is still helpful for the existing
implementations of TargetParams to keep FromStr, so they can still do
their parsing.

Signed-off-by: mulhern <amulhern@redhat.com>