Have LinearDev methods take a LinearDevTargetTable argument
Get rid of Segments entirely.
Expose LinearDevTable stuff in the public interface.

This change allows directly setting the table in the LinearDevTarget, to
allow both flakey and linear. It takes away the default organization of
the segments passed, which is contiguous, and leaves the responsibilty
for that arrangement up to the client.

Signed-off-by: mulhern <amulhern@redhat.com>