Remove support for initial broken DM event support
Support for the fixed support in DM minor version 37+ is now universal
enough that we can drop support for the broken initial version that
briefly existed in 4.13.

Why now? I think we could have removed it earlier, but the impetus was
updating the docs for how to use DM events. It's simplest to just claim
and document support starting with minor 37.

Signed-off-by: Andy Grover <agrover@redhat.com>