Remove DM struct
The sole purpose of the DM struct was to hold an open File for the device mapper
control file '/dev/mapper/control'.  With its existing use in shared.rs as immutable,
going forward we would be limited to what methods could be added to it.
With the file being stored in the structure it's open lifetime was tied to the
entire duration of the instance (effectively being open at all times).  This
limits the implementation on what it can and cannot do with the mapper control
file.  This change should give us more flexibility going forward and reduce the
need to pass the DM structure around virtually everywhere.

Signed-off-by: Tony Asleson <tasleson@redhat.com>