Add DmOption parameter to device_create
Current best approach for getting existing user space udev rules
to do what we need.  To allow Stratis to be used for root fs and for
allowing Stratis backed XFS fs mounted in fstab.  This works by
setting a cookie value when we resume the newly created thin
device which is utilized for newly created FS.

Signed-off-by: Tony Asleson <tasleson@redhat.com>