Fix enospc when testing on system  with small tmpfs
Previously get_devices was creating a sparse file by opening the file,
seeking and writing to the end.  Then it was converted to using fallocate
which was creating a non-sparse file.  Fallocate can make an existing file
sparse or add to an existing file as sparse, but for simplying creating
a sparse file to start with ftruncate is the simplest approach.

Signed-off-by: Tony Asleson <tasleson@redhat.com>