Change i8 fields in Struct_dm_ioctl to be u8
When we originally ran rust-bindgen, it translated char arrays in
dm-ioctl.h as libc::c_char, which is typed as i8. This is aggravating,
making us first cast to u8 arrays, which is more natural for their use as
fixed-size buffers for strings.

Instead, just change our struct definition to [u8].

Signed-off-by: Andy Grover <agrover@redhat.com>