Call out kernel bug if fields are not null-terminated
Usually the right thing to do is to assume the kernel has no bugs,
otherwise you go crazy, checking invariants. If we did that here, then we
could just call expect() instead of providing a default value. However,
this is for Debug, so it makes sense to try very hard to not panic, since
this could be what developers are trying to use to find a problem.

Signed-off-by: Andy Grover <agrover@redhat.com>