Do not implement Error and do explicitly implement Fail
The blanket implementation of Fail does not provide the backtrace, even
though the context has it. Implementing Error causes the blanket
implementation of Fail to be implemented, making it impossible for the
Fail trait to be implemented locally. We want the backtrace, so with this
failure approach, we have only one choice.

Signed-off-by: mulhern <amulhern@redhat.com>