Add an omnibus macro, range!, to gather up all easily gathered macros
It's less repetitive that way.

Change the definition of checked_add! so it can be used in range!

Define some subsidiary macros that group rem, div, and mul operations

Comment them to indicate that they are complete, i.e., there are no other
ways to define these operations on units and numbers.

Of course, there are other dimensions to be explored, like checked, saturating,
and assigning operators. Leave all that out of the discussion at this time.

Signed-off-by: mulhern <amulhern@redhat.com>