Turn str_check function into a macro and put it in id_macros
No longer pass as an argument to str_id macro. I think str_check was defined
originally as a function because we were novice macro users. That function
was always passed for construction of the type, it was used exclusively for
that purpose, and it seems unlikely that it will need to be varied.

Have str_check just evaluate to an Option<String> rather than a Result.
Promote the Option to the error in str_id. Removes duplication and
isolates the use of the DmResult type to two places.

Signed-off-by: mulhern <amulhern@redhat.com>