Remove the sudo commands in sudo-test Makefile target
It seems more reasonable to require the caller to do the sudoing.

Call the script directly for the tests that require sudo in Travis config.
These tests require sudo, because Travis must be root to create the
loopbacked devices. Make sure that the PATH includes the location where
Travis installs the Rust tools.

Signed-off-by: mulhern <amulhern@redhat.com>