Add list_test_devices method
This method lists devices whose names end with the the test device
suffix. In this PR, this method is called within our tests in lieu
of the list_devices method, which contrarily does not concern
itself with whether or not the devices it lists are test devices.
As a result, the possible existence of non-test devices no longer
interferes with test correctness.