Drop "cargo expand" target and Travis task
It wasn't so much CI as a quick-and-easy way to get a look at the results
of macro expansion for certain parts of devicemapper implementation.

It happens to be failing now, because rustfmt isn't currently available on
nightly.

There's a fix for that, using the flag "--component complete", but since
I'm pretty sure nobody uses the task anymore, it seems better to get rid of
it, for the goal of less maintenance.

Also, the Rust playground has added a macro expansion facility in the
meantime, which is quick and handy.

Signed-off-by: mulhern <amulhern@redhat.com>