Get rid of badges in the README
The first one would need to be updated for the move to .com instead of .org
and would also need to be updated if we started changing the default branch
to not master, and can be misleading for other reasons.

For some reason Fedora packaging complains about the second URL in the
crates.io badge.

Less trouble to just get rid of them and makes the README easier on the eyes
when reading as text.

Signed-off-by: mulhern <amulhern@redhat.com>