Device mapper control path for Android
On Android systems the dm control file is located at /dev/device-mapper.
The reason why this differs from "normal" Linux distributions is unclear
but since the path is hardcoded at many places, changing it or creating
a new chardev ad /dev/mapper/control is not an option. It is unlikely
that Android will change the path because this would break an enoumous
amount of systems.

Add a path definition when building the crate for target_os = "android".