Add options to DM::table_load
When loading a verity table the device *must* be a readonly. The dm
driver requires the flags for the DM_TABLE_LOAD_CMD ioctl to include
DM_READONLY. Add an argument to `core::dm::table_load`, that enables to
pass DmFlags::DM_READONLY to the DM_TABLE_LOAD_CMD operation.