Remove error-chain dependency in source
Store io errors as strings as they are not cloneable at this time.

Do not implement From for any error; be explicit to avoid accidental
conversion of some IO errors in the non-core part.

Put dependent errors in a Box, as well as DeviceInfo value. For Strings
and Pathbufs, just include them in the error.

Signed-off-by: mulhern <amulhern@redhat.com>