Introduce conditional compilation
Base the conditional compilation on a MAXIMUM_KERNEL_VERSION environment
variable.

Make the list of versions sparse. Here there is only one version
specified, the one that introduced the ARM_POLL ioctl number.

Specify that version or above and the arm_poll method is compiled in,
below it is not.

Signed-off-by: mulhern <amulhern@redhat.com>