Add task to print "dmsetup --version" weekly
This is really to help testing on stratisd. We need to make a special
test release of devcicemapper, with a hard-coded low driver version to
run our stratisd tests on containers.

The output of this will let us know what is the highest hard-coded
version that can be expected to work.

Signed-off-by: mulhern <amulhern@redhat.com>