Add core::dm_udev_sync sub-module
Add a new sub-module to encapsulate creating, destroying, and
waiting on device-mapper Udev synchronisation semaphores.

The module contains the low-level wrappers to the unsafe SysV semaphore
bindings provided by the libc crate as well as higher level UdevSync
interface for use by other sub-modules:

  UdevSync::begin()
  UdevSync::end()
  UdevSync::cancel()

This provides the necessary primitives to implement the ioctl-issuing
side of the DM cookie protocol. Completion of DM cookie semaphores is
not implemented as this is always carried out by the dmsetup program via
udev rule callbacks.

This simplifies integrating udev synchronization with the other
DM methods and avoids the need to expose the low-level semaphore
operations outside the dm_udev_sync module.

Since it is possible for the kernel to be compiled without semaphore
support or for the administrator to dynamically disable their use via
sysctl, runtime semaphore limits are checked in a new helper function,
dm_udev_sync::sysv_sem_supported().

Testing the maxid value returned by semctl(0, 0, SEM_INFO, ...) as is
done by libdevmapper is not sufficient to determine whether semaphores
can be used. It is possible for the administrator to disable them
completely via sysctl without causing the maxid value to be less than
zero:

  # echo 0 0 0 0 > /proc/sys/kernel/sem

Check that the following values in the returned seminfo struct are
greater than zero:

  semmsl - Maximum number of semaphores in a set
  semmni - Maximum number of semaphore sets
  semmns - Maximum number of semaphores in all semaphore sets

Additionally log a warning if the value of any of these settings is less
than 1000 which is an exceptional situation based on typical modern
distribution defaults for these settings.

Related: #774