Add core::dm_ioctl::ioctl_to_version()
Add a function to the dm_ioctl sub-module to map a device-mapper
ioctl command to a 3-tuple representing the minimum required kernel
interface version for that command.

This is based on the _cmd_data_v4 array defined in libdevmapper and
should be updated when adding new ioctl commands exposed by upstream
device-mapper changes.