Set per-command ioctl version in DM::do_ioctl()
Move initialization of the header version fields from to_ioctl_hdr()
into do_ioctl() and use ioctl_to_version() to set a per-command ioctl
interface version.

This allows commands to be issued as long as the kernel supports the
minimum interface version needed by a specific command, rather than
always expecting the kernel to provide the exact interface version that
devicemapper-rs-sys was compiled with.

For example, prior to this change if the running kernel supports DM
interface version 4.45.0 but devicemapper-rs-sys was compiled with
version 4.46.0 all ioctl commands will fail (since 4.45.0 < 4.46.0).