Fix failures due to udev reads in cachedev::tests::loop_test_minimal_cache_dev
Allowing udev time to interact with the devices invalidates some of the
assumptions in loop_test_minimal_cache_dev regarding the state of the
cache since udev probes will issue reads to the device to determine its
content, meaning that the cache is no longer completely pristine.

Remove the following assertions:

    assert_eq!(usage.used_cache, DataBlocks(0));
    assert_eq!(performance.read_hits, 0);
    assert_eq!(performance.read_misses, 0);
    assert_eq!(performance.promotions, 0);

The alternative would be to expose the DmOptions used to create the
cache device stack to users of the CacheDev, allowing device-mapper
udev flags to be used to suppress udev rule processing however this
would be a quite invasive change solely for the use of the test suite.