Do not use feature_args for thinpool equality
They are all instructions for how to behave, and could change for
reasons having to do with the environment. None of them indicate that the thin
pool device itself has changed, and so none of them should be included
in the equality comparison.

The motivation for this is that in some cases a pool will fail to be
properly started if the pool was created and fully allocated under a pre
3.1 stratisd release and then the running version of stratisd is
replaced with a newer version without the previously created pool being torn
down. In that case, the newer version of stratisd will calculate that
"error_if_no_space" should be a feature arg, but the older version will
not have that feature arg, and so the new version of stratisd will
consider the pool stopped, because it can't idempotently set up the
thin pool. This will happen even though really the pool is up and fully
functional.

The workaround is to tear down the thin pool device and everything on
top of it. Then the pool will start by itself or, if start/stop
functionality is available can be started with the start command.

Signed-off-by: mulhern <amulhern@redhat.com>