Add 'create_options' to device new() and setup() methods
Add the ability for callers to specify the DmOptions value to use
when creating or setting up lineardev, thindev, thinpooldev, and
cachedev.

The DmOptions value is wrapped in an Option which if set to None
gives the previous devicemapper-rs default behaviour.

This allows clients of the crate to opt in to controlling device
visibility and other DmOptions/DmUdevFlags behaviour when creating
or setting up devices.