Downgrade DM debugging messages to trace level
Commit 66e1fdf2 added logging to the DM module using the debug!
macro.

In practice these messages can be very noisy, especially for e.g.
in the Stratis test suite (which calls the status() method on thin pool
devices). This leads to lots of messages like the following:

[2023-11-06T16:01:05Z DEBUG devicemapper::core::dm] Retrieving table status for stratis-1-private-50db9929d265403b820bb155c8f5ae89-thinpool-pool

Running the Stratis test suite results in hundreds of these messages
being logged at debug level.

Downgrade the logging calls in the DM module from debug! to trace!
to quiet this down.